tools/confluence-migration/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
markdownify>=0.11.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
anthropic>=0.25.0   # Claude fallback classifier
python-slugify>=8.0.0
rich>=13.0.0        # terminal output, stage gate prompts
